backend/requirements.txt
boto3>=1.28.0
botocore>=1.31.0
PyJWT>=2.8.0
